requirements.txt
tensorboard==1.15.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.15.0
tensorflow-estimator==1.15.1
matplotlib==3.1.0
numpy==1.18.1
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
tqdm==4.47.0
Pillow==7.2.0
